// api/generated/model/getRecipesOrderBy.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Cheffy API
 * A Cheffy API é o backend central de uma aplicação de receitas, projetado para gerenciar catálogo culinário, categorias, tags, ingredientes, informações nutricionais, favoritos de usuário, autenticação via Better Auth, assinatura de upload no Cloudinary e um assistente gastronômico integrado à Groq.
 * OpenAPI spec version: 1.0.0
 */

export type GetRecipesOrderBy = typeof GetRecipesOrderBy[keyof typeof GetRecipesOrderBy];


export const GetRecipesOrderBy = {
  newest: 'newest',
  oldest: 'oldest',
} as const;